feat: commplete projects page
