+++ app/assets/css/main.postcss
@@ -18,6 +18,13 @@
   .gradient_text {
     @apply text-transparent bg-clip-text bg-gradient-to-t from-violet-950 to-violet-400 dark:from-zinc-400 dark:to-zinc-100;
   }
+
+  .entering { animation: 1.2s entering ease-out; }
+
+  @keyframes entering {
+    0% { clip-path: inset(0 0 100% 0); }
+    100% { clip-path: inset(0); }
+  }
 }
 
 html {
